make confconsole fault tolerant by handling exceptions in a user friendly way
Rational: errors/bugs are inevitable and we should attempt to minimize their
impact as much as possible. Exceptions shouldn't terminate the confconsole, as
it is very unfriendly to new users when the confconsole suddenly crashes.
Instead what we should do is intercept the exception, provide a useful
traceback that he can submit to us so we can squash the bug and return the user
to the previous dialog.

If we had this fault tolerance in place previously, it would have mitigated the
impact of the bug I previously discovered with the broken static IP
configuration screen and made it easier for that VMWare user to report his
problem.